md5 patch: package version must be decoded to use it
Whole uri comes encoded (urlencoded). Filename must NOT be decoded because
plain aptitude do not decode it when download and install it. Therefore, we
will have ugly named packages at /var/cache/apt/archives but is the standard
behavior.
But package version must be decoded, otherways package=version calls will
not work.
Notice: package version admit '+', so this character must not be translated.