add cleanup for mirrorlist and refactor TMP_ vars
The new mirrorlist variable is $MIRRORLIST.

The cleanup_dllist function is now refactored to accept some arguments.
The refactoring was intended for MIRRORLIST, but it then occurred to me
that it was not meant for persistence anyway. So now it gets
insta-deleted without need for a cleanup function.

TMP_ variables previously used a pretty maintaience-heavy way to be
assigned. I changed it to use the "nameref" magic in bash. Basically
these variables are read/write proxies to some other variables.