perf: utilize numfmt's cut-style formatting (2.06x speedup)
This is a follow-up of dc4ef45b ("perf: speed up (>3x) get_uris,
display_downloadfile with bash builtin"). In that patch, I left the
calculation of file size unchanged. It turns out that it is the severest
performance hot spot remaining.

The patch utilizes numfmt's cut-style formatting (--field=N) so that we
no longer need to convert file sizes into human-readable forms
repeatedly while parsing uris_full. This eliminates a lot of overheads
of subshell and piping, which consist of expensive fork()s and
execve()s.

A brief test (excluding the overhead of apt --print-uris) shows that the
speedup is 2.06x (5.86/2.85). This is done by executing the following:
$ apt -y --print-uris dist-upgrade | tee test-data | wc -l
3347
$ grep -E "^'(http(s|)|(s|)ftp)://" test-data | wc -l
1659
$ echo $'#!/bin/sh\nexec cat "test-data"' >apt && chmod +x apt
$ sudo PATH="$PWD:$PATH" APT_FAST_TIMEOUT=0 time apt-fast upgrade

Before:
5.20user 1.32system 0:05.86elapsed 111%CPU (0avgtext+0avgdata 12380maxresident)k
0inputs+0outputs (0major+1356629minor)pagefaults 0swaps
After:
2.20user 0.76system 0:02.85elapsed 103%CPU (0avgtext+0avgdata 12436maxresident)k
0inputs+0outputs (0major+616493minor)pagefaults 0swaps