perf: utilize bash operator += (1.28x speedup)
The patch utilizes bash operator += to speed up the appending to
DOWNLOAD_DISPLAY and the accumulation to DOWNLOAD_SIZE. This means the
modification can be done in place, eliminating the overhead of template
substitution and variable reassignment.

A brief test (excluding the overhead of apt --print-uris) shows that the
speedup is 1.28x (2.83/2.22).

Before:
2.20user 0.76system 0:02.85elapsed 103%CPU (0avgtext+0avgdata 12436maxresident)k
0inputs+0outputs (0major+616493minor)pagefaults 0swaps
After:
1.73user 0.60system 0:02.22elapsed 105%CPU (0avgtext+0avgdata 12300maxresident)k
0inputs+0outputs (0major+487300minor)pagefaults 0swaps