perf: inline urldecode() and skip it when unneeded (1.37x speedup)
The only user of urldecode() is get_uris() and the former is only called
in one place in a loop. It turns out that it is one of the severest
performance hot spots remaining (the other one is get_auth(), but
inlining it would hurt the readability).

The patch inlines urldecode() into the only place calling it, while
conditionally skipping it when the file name does not contain any
urlencoded string. When a file name needs to be decoded, the inlining
eliminates the overheads of subshell and function calls, which consist
of an expensive fork(). When a file name doesn't need to be decoded
(this is very likely, in my case, 73.8%), skipping decoding further
eliminates an additional subshell to call printf.

A brief test (excluding the overhead of apt --print-uris) shows that the
speedup is 1.37x (2.22/1.62).

Before:
1.73user 0.60system 0:02.22elapsed 105%CPU (0avgtext+0avgdata 12300maxresident)k
0inputs+0outputs (0major+487300minor)pagefaults 0swaps
After:
1.25user 0.45system 0:01.62elapsed 105%CPU (0avgtext+0avgdata 12328maxresident)k
0inputs+0outputs (0major+331489minor)pagefaults 0swaps